Better icon and style fallback and messaging
Missing icons will now fall back to “question-circle” available in the FA free set, and an error message will be printed.

If no style is specified and “regular” is not available, the first available style will be used instead, but silently.

If a style is specified but unavailable, same behavior as above, but an error message will be printed to log.